requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
anthropic==0.40.0
google-genai==1.14.0
beautifulsoup4==4.12.3
